# backend/src/main/resources/openapi/openapi.yaml
openapi: "3.1.0"
info:
  title: Duty Tracker API
  description: On-Call Hours Tracker — REST API for managing on-call periods, incidents, compensation rates, and engineer profiles
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/v1/compensation-rates:
    get:
      operationId: getAllCompensationRates
      summary: Get compensation rate table
      description: Retrieve all compensation rates
      tags:
        - Compensation Rates
      responses:
        '200':
          description: Compensation rates retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompensationRateTableResponse'

    post:
      operationId: createCompensationRate
      summary: Create compensation rate
      description: Create a new compensation rate with percentage and label
      tags:
        - Compensation Rates
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCompensationRateRequest'
      responses:
        '201':
          description: Compensation rate created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompensationRateResponse'
        '400':
          description: Invalid rate data
  /api/v1/compensation-rates/{id}:
    put:
      operationId: updateCompensationRate
      summary: Update compensation rate
      description: Update the percentage and label of a compensation rate
      tags:
        - Compensation Rates
      parameters:
        - name: id
          in: path
          required: true
          description: Compensation rate ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCompensationRateRequest'
      responses:
        '200':
          description: Compensation rate updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompensationRateResponse'
        '400':
          description: Invalid rate data
        '404':
          description: Compensation rate not found

    delete:
      operationId: deleteCompensationRate
      summary: Delete compensation rate
      description: Remove a compensation rate from the system
      tags:
        - Compensation Rates
      parameters:
        - name: id
          in: path
          required: true
          description: Compensation rate ID
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Compensation rate deleted successfully
        '404':
          description: Compensation rate not found

  # ── On-Call Periods ───────────────────────────────────────────────────────
...
  /api/v1/incidents:
    post:
      operationId: logIncident
      summary: Log a new incident
      description: Create a new incident entry for overtime tracking
      tags:
        - Incidents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogIncidentRequest'
      responses:
        '201':
          description: Incident created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentResponse'
        '400':
          description: Invalid incident data

    get:
      operationId: listIncidents
      summary: List incidents
      description: Retrieve all incidents, optionally filtered by on-call period
      tags:
        - Incidents
      parameters:
        - name: onCallPeriodId
          in: query
          required: false
          description: On-call period ID to filter incidents (optional)
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Incidents retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentListResponse'
  /api/v1/incidents/{id}:
    get:
      operationId: getIncident
      summary: Get incident by ID
      description: Retrieve details of a specific incident
      tags:
        - Incidents
      parameters:
        - name: id
          in: path
          required: true
          description: Incident ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Incident retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentResponse'
        '404':
          description: Incident not found

    put:
      operationId: updateIncident
      summary: Update incident
      description: Update the date and time of an incident
      tags:
        - Incidents
      parameters:
        - name: id
          in: path
          required: true
          description: Incident ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateIncidentBody'
      responses:
        '200':
          description: Incident updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentResponse'
        '400':
          description: Invalid incident data
        '404':
          description: Incident not found

    delete:
      operationId: deleteIncident
      summary: Delete incident
      description: Remove an incident from the system
      tags:
        - Incidents
      parameters:
        - name: id
          in: path
          required: true
          description: Incident ID
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Incident deleted successfully
        '404':
          description: Incident not found
  /api/v1/incidents/{id}/calculate:
    post:
      operationId: calculateIncidentOvertime
      summary: Calculate overtime entries
      description: Generate overtime entries for a specific incident
      tags:
        - Incidents
      parameters:
        - name: id
          in: path
          required: true
          description: Incident ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Overtime entries calculated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OvertimeEntriesResponse'
        '404':
          description: Incident not found

  # ── Holidays ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    CreateCompensationRateRequest:
      type: object
      required:
        - overtimeDayType
        - label
        - timeFrom
        - timeTo
        - percentage
      properties:
        overtimeDayType:
          $ref: '#/components/schemas/OvertimeDayType'
        label:
          type: string
        timeFrom:
          type: string
          format: time
          example: "09:00:00"
        timeTo:
          type: string
          format: time
          example: "17:00:00"
        percentage:
          type: number
          format: decimal
    UpdateCompensationRateRequest:
      type: object
      required:
        - percentage
        - label
      properties:
        percentage:
          type: number
          format: decimal
        label:
          type: string
    CompensationRateResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        rateCategory:
          $ref: '#/components/schemas/RateCategory'
        overtimeDayType:
          $ref: '#/components/schemas/OvertimeDayType'
        label:
          type: string
        timeFrom:
          type: string
          format: time
        timeTo:
          type: string
          format: time
        percentage:
          type: number
          format: decimal
    CompensationRateTableResponse:
      type: object
      properties:
        rates:
          type: array
          items:
            $ref: '#/components/schemas/CompensationRateResponse'

    # ── On-Call Periods ───────────────────────────────────────────────────
...
    LogIncidentRequest:
      type: object
      required:
        - onCallPeriodId
      properties:
        onCallPeriodId:
          type: integer
          format: int64
        name:
          type: string
        startDateTime:
          type: string
          format: date-time
        endDateTime:
          type: string
          format: date-time
    UpdateIncidentBody:
      type: object
      properties:
        name:
          type: string
        startDateTime:
          type: string
          format: date-time
        endDateTime:
          type: string
          format: date-time
    IncidentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        onCallPeriodId:
          type: integer
          format: int64
        name:
          type: string
        startDateTime:
          type: string
          format: date-time
        endDateTime:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    IncidentListResponse:
      type: object
      properties:
        incidents:
          type: array
          items:
            $ref: '#/components/schemas/IncidentResponse'
    OvertimeEntriesResponse:
      type: object
      properties:
        incidentId:
          type: integer
          format: int64
        entries:
          type: array
          items:
            $ref: '#/components/schemas/OvertimeEntryResponse'
    OvertimeEntryResponse:
      type: object
      properties:
        incidentId:
          type: integer
          format: int64
        overtimeHours:
          type: number
          format: decimal
        allowanceHours:
          type: number
          format: decimal
        allowancePercentage:
          type: number
          format: decimal
        date:
          type: string
          format: date
        timeFrom:
          type: string
          format: time
        timeTo:
          type: string
          format: time
        isAllowanceEntry:
          type: boolean

    # ── Holidays ──────────────────────────────────────────────────────────
...
    OvertimeDayType:
      type: string
      enum:
        - WEEKDAY
        - SATURDAY
        - SUNDAY_HOLIDAY
    RateCategory:
      type: string
      enum:
        - ONCALL_WEEKDAY_SATURDAY
        - ONCALL_SUNDAY_HOLIDAY
        - OVERTIME_BASE
        - OVERTIME_ALLOWANCE
...